refactor(test): split useBookmarkPage.test.ts into functional units and improve infrastructure (#569)
